Split BuilderClassTest into different tests
